feat: 🎸 convert constants based on locale
