+++ package.json
@@ -25,6 +25,9 @@
     "emoji-picker-react": "^3.5.1",
     "framer-motion": "^4.1.17",
     "hls.js": "^1.1.5",
+    "i18next-chained-backend": "^3.0.2",
+    "i18next-http-backend": "^1.4.0",
+    "i18next-localstorage-backend": "^3.1.3",
     "next": "^12.0.9",
     "next-i18next": "^11.0.0",
     "next-pwa": "^5.4.4",
