+++ public/locales/en/common.json
@@ -1,1 +1,1 @@
-{"genres":"Genres","language":"English","most_favourite":"All time favorite","most_favourite_season":"Favorite this season","most_popular":"All time popular","most_popular_season":"Popular this season","newly_added":"Newly updated","recent_comments":"Recent comments"}
+{"chapter":"Chapter","episode":"Episode","genres":"Genres","most_favourite":"All time favorite","most_favourite_season":"Favorite this season","most_popular":"All time popular","most_popular_season":"Popular this season","newly_added":"Newly updated","recent_comments":"Recent comments"}
