+++ public/locales/vi/common.json
@@ -1,1 +1,1 @@
-{"genres":"Thể loại","language":"Tiếng Việt","most_favourite":"Được yêu thích","most_favourite_season":"Được yêu thích mùa này","most_popular":"Nổi bật nhất","most_popular_season":"Nổi bật mùa này","newly_added":"Mới cập nhật","recent_comments":"Bình luận gần đây"}
+{"chapter":"Chương","episode":"Tập","genres":"Thể loại","most_favourite":"Được yêu thích","most_favourite_season":"Được yêu thích mùa này","most_popular":"Nổi bật nhất","most_popular_season":"Nổi bật mùa này","newly_added":"Mới cập nhật","recent_comments":"Bình luận gần đây"}
