+++ src/components/shared/LanguageSwitcher.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from "next-i18next";
 import { useRouter } from "next/router";
 import nookies from "nookies";
-import React, { useMemo } from "react";
+import React from "react";
 import { MdOutlineLanguage } from "react-icons/md";
 import Popup from "./Popup";
 
@@ -15,7 +15,7 @@ const locales = [
 
 const LanguageSwitcher = () => {
   const router = useRouter();
-  const { i18n, t } = useTranslation();
+  const { i18n } = useTranslation();
 
   const handleChangeLanguage = (lang: string) => () => {
     if (router.locale === lang) return;
@@ -27,15 +27,8 @@ const LanguageSwitcher = () => {
       shallow: true,
     });
 
-    i18n.changeLanguage(lang);
-
     nookies.set(null, "NEXT_LOCALE", lang, { path: "/" });
   };
-
-  const languages = useMemo(
-    () => Object.keys(i18n.store.data),
-    [i18n.store.data]
-  );
 
   return (
     <Popup
@@ -46,14 +39,14 @@ const LanguageSwitcher = () => {
       }
     >
       <ul className="space-y-1">
-        {languages.map((lang) => (
+        {locales.map(({ locale, name }) => (
           <li
             className="cursor-pointer transition duration-300 hover:text-primary-300"
-            onClick={handleChangeLanguage(lang)}
-            key={lang}
-            title={lang}
+            onClick={handleChangeLanguage(locale)}
+            key={locale}
+            title={locale}
           >
-            {t("language", { lng: lang })}
+            {name}
           </li>
         ))}
       </ul>
